refactor(frontend): Add solution tasks component
