+++ frontend/src/app/assignment/pages/solution/solution.component.html
@@ -37,28 +37,7 @@
         Solution & Tasks
       </a>
       <ng-template ngbNavContent>
-        <div class="row">
-          <div class="col-md-6" *ngIf="solution && solution.solution">
-            <div class="form position-sticky" style="top: 4rem">
-              <div class="form-group">
-                <label for="solutionInput">Solution</label>
-                <app-scenario-codemirror
-                  id="solutionInput"
-                  [content]="solution ? solution.solution : 'Loading...'"
-                  [markers]="markers"
-                  [readOnly]="true"
-                ></app-scenario-codemirror>
-              </div>
-              <p class="text-center">
-                Submitted <b>{{solution.timestamp | date:'medium'}}</b>.
-              </p>
-            </div>
-          </div>
-          <div class="col-12" [class.col-md-6]="solution && solution.solution">
-            Tasks
-            <app-task-list [tasks]="assignment?.tasks" [evaluations]="evaluations" [points]="points"></app-task-list>
-          </div>
-        </div>
+        <app-solution-tasks></app-solution-tasks>
       </ng-template>
     </li>
     <li ngbNavItem="info">
