+++ frontend/src/app/assignment/pages/solution/tasks/tasks.component.html
@@ -1,0 +1,22 @@
+<div class="row">
+  <div class="col-md-6" *ngIf="solution && solution.solution">
+    <div class="form position-sticky" style="top: 4rem">
+      <div class="form-group">
+        <label for="solutionInput">Solution</label>
+        <app-scenario-codemirror
+          id="solutionInput"
+          [content]="solution ? solution.solution : 'Loading...'"
+          [markers]="markers"
+          [readOnly]="true"
+        ></app-scenario-codemirror>
+      </div>
+      <p class="text-center">
+        Submitted <b>{{solution.timestamp | date:'medium'}}</b>.
+      </p>
+    </div>
+  </div>
+  <div class="col-12" [class.col-md-6]="solution && solution.solution">
+    Tasks
+    <app-task-list [tasks]="assignment?.tasks" [evaluations]="evaluations" [points]="points"></app-task-list>
+  </div>
+</div>
